Develop (#23)
* Refactor computer nodes, use entrypoints

* add takeoff test, ros node communcation is broken

* takeoff from ros, entrypoint for gzserver

* Autodetect mavlink partner

* Double takeoff

* Use iris_1 and remove computer 2 stuff

* Add fluid and ascend_msgs as submodules

* Feature/auto build and upload (#21)

* Add mock plugin for testing

* Update README.md

* Add image pushing to makefile

Note: this does not ensure auto pushing. I'd like to make the build server autobuild and update docker-images with appropriate tags, but had problems compiling sitl_gazebo on the server. It works fine locally but on the server it attempts to create a directory /.gazebo which it isn't allowed to do, and also shouldn't do since it is outside of the jenkins workspace.

* Use single iris

* Add release tagging to uploaded images

* Add release tagging to uploaded images